chore: code tidying
