+++ data/db/files_related_morphs.json
@@ -442,7 +442,7 @@
   {
     "field": "Image",
     "file_id": 79,
-    "id": 56,
+    "id": 73,
     "order": 1,
     "related_id": 2,
     "related_type": "home.mission-statement"
