+++ data/db/sqlite_sequence.json
@@ -17,7 +17,7 @@
   },
   {
     "name": "blog_posts",
-    "seq": 17
+    "seq": 19
   },
   {
     "name": "blog_tags",
@@ -69,7 +69,7 @@
   },
   {
     "name": "components_home_mission_statements_components",
-    "seq": 10
+    "seq": 11
   },
   {
     "name": "components_home_project_summaries",
@@ -93,7 +93,7 @@
   },
   {
     "name": "donate_contents_components",
-    "seq": 6
+    "seq": 7
   },
   {
     "name": "faqs",
