+++ frontend/pages/home.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import type { GetStaticPropsContext, InferGetStaticPropsType } from "next";
 import Head from "next/head";
-import Link from "next/link";
 
 import { serverQuery } from "lib/graphql";
 import type { HomeContentQuery, ProjectType } from "../graphql/generated";
@@ -13,6 +12,7 @@ import { ProjectSummaryComponent } from "components/pages/home/ProjectSummary";
 import { GetInvolvedComponent } from "components/pages/home/getInvolved";
 import { ImpactSectionComponent } from "components/pages/home/ImpactSection";
 import { BlogCardComponent } from "components/pages/blog-post/blogCard";
+import { DonateSummary } from "components/pages/home/donateSummary";
 
 export const getStaticProps = async ({}: GetStaticPropsContext) => {
   const res = await serverQuery<HomeContentQuery>(HomeContentDocument);
@@ -27,17 +27,6 @@ export const getStaticProps = async ({}: GetStaticPropsContext) => {
     },
   };
 };
-
-const DonateHeartSVG = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="2"
-      d="M4.318 6.318a4.5 4.5 0 000 6.364L12 20.364l7.682-7.682a4.5 4.5 0 00-6.364-6.364L12 7.636l-1.318-1.318a4.5 4.5 0 00-6.364 0z"
-    />
-  </svg>
-);
 
 const HomePage = ({ content, blogs, projects }: InferGetStaticPropsType<typeof getStaticProps>) => (
   <>
@@ -71,20 +60,10 @@ const HomePage = ({ content, blogs, projects }: InferGetStaticPropsType<typeof g
           ))}
         </div>
       </PageSection>
-      <PageSection fullwidth className="bg-accent text-white py-0">
-        <div className="flex justify-center items-center">
-          <h3 className="text-center">You can help us make a difference</h3>
-          <Link href="/donate" className="ml-8">
-            <button className="btn btn-accent text-white border-white border-1 hover:!border-white">
-              <DonateHeartSVG />
-              <span className="ml-1">Donate</span>
-            </button>
-          </Link>
-        </div>
+      <PageSection fullwidth className="bg-accent text-white py-4">
+        <DonateSummary />
       </PageSection>
-      {/* <PageSection fullwidth className="bg-base-200 py-16">
-        (TODO - donate buttons)
-      </PageSection> */}
+      {/* <PageSection fullwidth className="bg-base-200 py-16"></PageSection> */}
     </>
   </>
 );
